backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
djangorestframework-camel-case>=1.4
Pillow>=10.0
python-dotenv>=1.0
gunicorn>=22.0
whitenoise>=6.6
psycopg[binary]>=3.1
pytest>=8.0
pytest-django>=4.8
dj-database-url>=2.2
django-filter>=24.2
celery>=5.3.0
redis>=5.0.0
boto3>=1.34.0
django-storages[s3]>=1.14.0
